fix bug in integration limits set better default plot offsets for reflection marks & difference curve.
git-svn-id: https://subversion.xray.aps.anl.gov/pyGSAS/trunk@2477 2b3fe19a-dc8f-cc80-9b5f-a4fb0fcdb5be

+++ GSASIIimgGUI.py
@@ -763,7 +763,7 @@ def UpdateImageControls(G2frame,data,masks,IntegrateOnly=False):
             G2plt.PlotExposedImage(G2frame,event=event)
             
         def OnFullIntegrate(event):
-            Lazm =int(G2frame.Lazim.GetValue())
+            Lazm = float(G2frame.Lazim.GetValue())
             if data['fullIntegrate']:
                 data['fullIntegrate'] = False
                 data['LRazimuth'] = [Lazm,Lazm+20.]
